Add player retention analysis graph (#2917)
Adds a graph to network and server pages that has options to:
- Draw graphs at different time resolutions
- Limit input data by time
- Group players by register date or join address
- Visualize retention in different ways
  - Time since registration date
  - Playtime
  - Date
  - Cumulative player gain
  - Percentage / Player count / Stacked player count

Any and all combinations are allowed which allows extensive analysis of player retention. Help sections attempt to make the data understandable and show examples.

Affects issues:
- Close #2159

+++ Plan/common/src/test/java/extension/FullSystemExtension.java
@@ -93,7 +93,7 @@ public class FullSystemExtension implements ParameterResolver, BeforeAllCallback
 
     @Override
     public void afterAll(ExtensionContext context) throws Exception {
-        deleteDirectory(tempDir);
+        if (tempDir != null) deleteDirectory(tempDir);
     }
 
     private void deleteDirectory(Path directory) throws IOException {
